Optimize PBC RSJK (#747)
* Optimize basis contraction in rsjk

* Optimize basis recontraction

* Update AFT get_k_kpts

* Fix bugs caused by unsynchronized threads in PBC create_tasks step

* Handle complex-valued density matrices in rsjk

* lint

* Unrolling pbc/rys_contract_k.cu

* Update unrolled_rys_k

* Optimize rys_contract_k

* Adjust omega estimation for rsjk

* Optimize aft_jk memory efficiency

* Use pre-allocated buffer for aft_jk.get_k

* bugfix

* Tune for very diffuse orbitals

* Fix aft_jk and update tests

* lint

* Update tests

* Tune basis set contraction

+++ gpu4pyscf/lib/pbc/rys_contract_jk_ip1.cu
@@ -161,7 +161,7 @@ while (1) {
 
         for (int task_id = sq_id; task_id < ntasks+sq_id; task_id += nsq_per_block) {
             __syncthreads();
-            int bas_kl = bas_kl_idx[task_id];
+            uint32_t bas_kl = bas_kl_idx[task_id];
             int ksh = bas_kl / nbas;
             int lsh = bas_kl % nbas;
             int _ksh = bas_mask_idx[ksh];
@@ -702,7 +702,7 @@ while (1) {
 
         for (int task_id = sq_id; task_id < ntasks+sq_id; task_id += nsq_per_block) {
             __syncthreads();
-            int bas_kl = bas_kl_idx[task_id];
+            uint32_t bas_kl = bas_kl_idx[task_id];
             int ksh = bas_kl / nbas;
             int lsh = bas_kl % nbas;
             int _ksh = bas_mask_idx[ksh];
@@ -749,7 +749,7 @@ while (1) {
             double v_lx = 0;
             double v_ly = 0;
             double v_lz = 0;
-            int nao2 = nao * nao;
+            size_t nao2 = nao * nao;
             double *dm_jk = dm + Ts_ij_lookup[cell_j+cell_k*nimgs] * nao2;
             double *dm_jl = dm + Ts_ij_lookup[cell_j+cell_l*nimgs] * nao2;
             double *dm_ki = dm + Ts_ij_lookup[cell_k             ] * nao2;
